reorganization (copy prop yet to be completely done)
